pyproject.toml
[project]
name = "toybridge"
version = "0.1.0"
description = "Reverse-engineer any BLE toy's protocol, then control it with AI (Claude / OpenClaw)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleak>=2.1.1",
    "mcp[cli]>=1.26.0",
    "pyobjc-framework-corebluetooth>=12.1",
]
